Update StageGenerator
# StageGenerator
#### Details:
The StageGenerator class now generates stages. For now the stagse are created with z-axis offset. Also there is a Level system now meaning each levels will have certain amount of stages. Just need to implement the height and side offsets for better look and feel. I have commented the variables and methods in the code.
***