Update StageGenerator.cs
# StageGenerator
#### Details:
This class now places obstacles in the game world. The fields and properties used for placing obstacles are found under the **[Header("Obstacle Properties")]**. Read the comments there to understand more about the fields and properties.

The logic used to add obstacles can be found in the method **void AddLinkPoint(BouncyStage)**. Read the comments there to understand more. May make a different method for the logic of the obstacles but the placement will be same.
***