Update BouncyStageBreakable.cs
# Default value at **Awake()** method
#### Details:
Setting the default value at the **Awake()** method because this will ensure that the default value has been set up before the **StageGenerator** class gets to call **ResetStage()** method.

Before this changed I was facing a bug where the **StageGenerator** class was reseting the breakable stage but after the call the **BouncyStageBreakable** was calling its **void Start()** method.
***