Created JumpRaceInput
# JumpRaceInput
#### Details:
This is the main input system of the game. It looks for touches on the screen and based on the logic given here updates the _xAxis value which is later used by the Player class to rotate the player.

Also gave a hint in the **void ResetXAxis()** on how to smooth the _xAxis value if needed. It is commented out.
***