Update Enemy.cs
# Random end stage position
#### Details:
The enemy now generates a random position within the end platform to land on. This is working fine. So far no enemies have clipped into each other but there is always a possibility for that. This was it gives a uniquness to the end platform landing.

For 100% no landing within each other then the random position generation should be made fixed and each enemies should ask for a fixed position to land on. It could be given in the EnemyGenerator.
***