Update FloorDetector.cs
# Parent BasicLineDetector
#### Details:
Changed the parent of the this class. Took the common functionalities to the parent class. Kept the old code in the comments. Will remove it later if the comment code is no longer needed. The shifting of the logic was necessary because FakeShadow uses the same functionality as this class but with minute differences.
***